requirements.txt
Flask
Werkzeug
nltk
gensim
pyvi
datasets
evaluate
scikit-learn
numpy
python-docx
pymupdf
pytesseract
pillow
